refactor: extract giftNoteService
