+++ src/main/java/com/twoCube/gifts/service/GiftService.java
@@ -7,7 +7,7 @@ import org.springframework.web.multipart.MultipartFile;
 import java.util.List;
 
 public interface GiftService {
-    Long createNote(NoteRequest noteRequest, Member member);
+//    Long createNote(NoteRequest noteRequest, Member member);
     FlowerResponse getRandomFlower(Member member);
     Long createFlower(FlowerRequest flowerRequest, Member member);
     List<PillListResponse> getPillList();
